+++ packages/core/react/use-chat.ts
@@ -61,7 +61,7 @@ export type UseChatHelpers = {
       | React.ChangeEvent<HTMLInputElement>
       | React.ChangeEvent<HTMLTextAreaElement>,
   ) => void;
-  /** Form submission handler to automatically reset input and append a user message  */
+  /** Form submission handler to automatically reset input and append a user message */
   handleSubmit: (
     e: React.FormEvent<HTMLFormElement>,
     chatRequestOptions?: ChatRequestOptions,
@@ -99,14 +99,20 @@ const getStreamedResponse = async (
 
   const constructedMessagesPayload = sendExtraMessageFields
     ? chatRequest.messages
-    : chatRequest.messages.map(({ role, content, name, function_call }) => ({
-        role,
-        content,
-        ...(name !== undefined && { name }),
-        ...(function_call !== undefined && {
-          function_call: function_call,
+    : chatRequest.messages.map(
+        ({ role, content, name, function_call, tool_calls, tool_call_id }) => ({
+          role,
+          content,
+          tool_call_id,
+          ...(name !== undefined && { name }),
+          ...(function_call !== undefined && {
+            function_call: function_call,
+          }),
+          ...(tool_calls !== undefined && {
+            tool_calls: tool_calls,
+          }),
         }),
-      }));
+      );
 
   if (typeof api !== 'string') {
     // In this case, we are handling a Server Action. No complex mode handling needed.
@@ -166,6 +172,12 @@ const getStreamedResponse = async (
       ...(chatRequest.function_call !== undefined && {
         function_call: chatRequest.function_call,
       }),
+      ...(chatRequest.tools !== undefined && {
+        tools: chatRequest.tools,
+      }),
+      ...(chatRequest.tool_choice !== undefined && {
+        tool_choice: chatRequest.tool_choice,
+      }),
     },
     credentials: extraMetadataRef.current.credentials,
     headers: {
@@ -196,6 +208,7 @@ export function useChat({
   initialInput = '',
   sendExtraMessageFields,
   experimental_onFunctionCall,
+  experimental_onToolCall,
   onResponse,
   onFinish,
   onError,
@@ -287,6 +300,7 @@ export function useChat({
               sendExtraMessageFields,
             ),
           experimental_onFunctionCall,
+          experimental_onToolCall,
           updateChatRequest: chatRequestParam => {
             chatRequest = chatRequestParam;
           },
@@ -323,6 +337,7 @@ export function useChat({
       streamData,
       sendExtraMessageFields,
       experimental_onFunctionCall,
+      experimental_onToolCall,
       messagesRef,
       abortControllerRef,
       generateId,
@@ -332,7 +347,14 @@ export function useChat({
   const append = useCallback(
     async (
       message: Message | CreateMessage,
-      { options, functions, function_call, data }: ChatRequestOptions = {},
+      {
+        options,
+        functions,
+        function_call,
+        tools,
+        tool_choice,
+        data,
+      }: ChatRequestOptions = {},
     ) => {
       if (!message.id) {
         message.id = generateId();
@@ -344,6 +366,8 @@ export function useChat({
         data,
         ...(functions !== undefined && { functions }),
         ...(function_call !== undefined && { function_call }),
+        ...(tools !== undefined && { tools }),
+        ...(tool_choice !== undefined && { tool_choice }),
       };
 
       return triggerRequest(chatRequest);
@@ -352,7 +376,13 @@ export function useChat({
   );
 
   const reload = useCallback(
-    async ({ options, functions, function_call }: ChatRequestOptions = {}) => {
+    async ({
+      options,
+      functions,
+      function_call,
+      tools,
+      tool_choice,
+    }: ChatRequestOptions = {}) => {
       if (messagesRef.current.length === 0) return null;
 
       // Remove last assistant message and retry last user message.
@@ -363,6 +393,8 @@ export function useChat({
           options,
           ...(functions !== undefined && { functions }),
           ...(function_call !== undefined && { function_call }),
+          ...(tools !== undefined && { tools }),
+          ...(tool_choice !== undefined && { tool_choice }),
         };
 
         return triggerRequest(chatRequest);
@@ -373,6 +405,8 @@ export function useChat({
         options,
         ...(functions !== undefined && { functions }),
         ...(function_call !== undefined && { function_call }),
+        ...(tools !== undefined && { tools }),
+        ...(tool_choice !== undefined && { tool_choice }),
       };
 
       return triggerRequest(chatRequest);